refactor(identity): split authentication responsibilities and implement session revocation
- Separate account, authentication and session services
- Move session creation from AuthenticationService to SessionService
- Introduce CreateSessionRequest and SessionResponse contracts
- Refactor login flow to delegate session creation
- Organize identity features into Account, Authentication and Session modules
- Separate entity and identity feature/controller structure
- Group token abstractions into Access and Refresh namespaces
- Implement session revoke and revoke-all operations
- Add revoke repository support using ExecuteUpdateAsync
- Add session revoke command handlers and endpoints
- Improve overall identity architecture and responsibility boundaries

+++ Core/DriverHub.Domain/Entities/Car.cs
@@ -11,7 +11,7 @@ public sealed class Car : Entity
     public int Km { get; set; }
     public string Transmission { get; set; } = string.Empty;
     public byte Seat { get; set; }
-    public byte Luggage { get; set; }
+    public int Luggage { get; set; }
     public string Fuel { get; set; } = string.Empty;
     public string BigImageUrl { get; set; } = string.Empty;
     public ICollection<CarFeature> CarFeatures { get; set; } = [];
